Misc. bug fixes and Debug plugin log feature. closes #17 closes #13
Added a log feature to the Debug plugin which shows a variable in the
browser after an AJAX request. Can be used with the following code:

    $this->esoTalk->plugins["Debug"]->log($variable1, $variable2, ...);

Note that this should only be used when debugging, because it will cause a
fatal error when the Debug plugin is not enabled.

Added/improved flood control when creating conversations.

Fixed various bugs including one in the formatter, Debug plugin, and with
unread post handling in conversations.

Signed-off-by: Toby <toby@esotalk.com>